Reworked ShowFilePopupMenuAction to have keyboard short-cut and work both on Parent dir and Search results
