+++ mucommander-core/src/main/java/com/mucommander/ui/action/ActionManager.java
@@ -201,7 +201,6 @@ public class ActionManager {
     	registerAction(new ShowRecentLocationsQLAction.Descriptor(),        ShowRecentLocationsQLAction::new);
     	registerAction(new ShowRootFoldersQLAction.Descriptor(), 			ShowRootFoldersQLAction::new);
     	registerAction(new ShowServerConnectionsAction.Descriptor(),        ShowServerConnectionsAction::new);
-        registerAction(new ShowTablePopupMenuAction.Descriptor(),           ShowTablePopupMenuAction::new);
     	registerAction(new ShowTabsQLAction.Descriptor(),					ShowTabsQLAction::new);
     	registerAction(new SortByDateAction.Descriptor(),             		SortByDateAction::new);
     	registerAction(new SortByExtensionAction.Descriptor(),              SortByExtensionAction::new);
